move Iterator to helpers, add hoc to List component
