+++ packages/elements/src/List/index.ts
@@ -1,3 +1,6 @@
 import component from './component'
+import withActiveState from './withActiveState'
+
+export { withActiveState }
 
 export default component
