+++ packages/elements/src/helpers/Iterator/index.ts
@@ -1,8 +1,5 @@
 import component from './component'
-import withActiveState from './withActiveState'
 import type { Props } from './component'
-
-export { withActiveState }
 
 export type { Props }
